+++ internal/core/lower/trigger_test.go
@@ -12,12 +12,13 @@ func TestTriggerClassification_SignalTrigger(t *testing.T) {
 	src := `
 		package test {
 			state M {
-				initial start;
+				entry; then start;
+				state start;
 				state waiting;
 				final done;
 				
 				succession first start then waiting;
-				transition waiting to done when powerOn;
+				transition first waiting when powerOn then done;
 			}
 		}
 	`
@@ -113,12 +114,13 @@ func TestTriggerClassification_ChangeEvent(t *testing.T) {
 	src := `
 		package test {
 			state M {
-				initial start;
+				entry; then start;
+				state start;
 				state waiting;
 				final done;
 				
 				succession first start then waiting;
-				transition waiting to done when x > 5;
+				transition first waiting when x > 5 then done;
 			}
 		}
 	`
@@ -188,12 +190,13 @@ func TestTriggerClassification_NilCompletion(t *testing.T) {
 	src := `
 		package test {
 			state M {
-				initial start;
+				entry; then start;
+				state start;
 				state waiting;
 				final done;
 				
 				succession first start then waiting;
-				transition waiting to done;
+				transition first waiting then done;
 			}
 		}
 	`
@@ -259,12 +262,13 @@ func TestTriggerClassification_CallTrigger(t *testing.T) {
 	src := `
 		package test {
 			state M {
-				initial start;
+				entry; then start;
+				state start;
 				state waiting;
 				final done;
 
 				succession first start then waiting;
-				transition waiting to done accept setSpeed(value);
+				transition first waiting accept setSpeed(value) then done;
 			}
 		}
 	`